backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-migrate
gunicorn
python-docx
PyPDF2
Flask-Limiter
pyjwt
werkzeug
